made tests pass with merge
